requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
PyJWT>=2.8.0
google-generativeai>=0.3.2
httpx>=0.25.2
aiofiles>=23.2.1
Pillow>=10.3.0
